Move ParticleBunch and LaserPulse
