+++ wake_t/__init__.py
@@ -2,6 +2,9 @@ from .__version__ import __version__
 from .beamline_elements import (PlasmaStage, PlasmaRamp, PlasmaLens, Drift,
                                 Dipole, Quadrupole, Sextupole, Beamline)
 from .physics_models.collective_effects.csr import set_csr_settings
+from .physics_models.laser.laser_pulse import LaserPulse
+from .particles.particle_bunch import ParticleBunch
 
 __all__ = ['__version__', 'PlasmaStage', 'PlasmaRamp', 'PlasmaLens', 'Drift',
-           'Dipole', 'Quadrupole', 'Sextupole', 'Beamline', 'set_csr_settings']
+           'Dipole', 'Quadrupole', 'Sextupole', 'Beamline', 'set_csr_settings',
+           'LaserPulse', 'ParticleBunch']
